Allow user to vertically scroll while toucing carousel
HammerJS was intercepting the vertical scroll events on the carousel.
The fix unfortunetely required a ton of code just to get it working on
android. I had to create a hacky solution by not ilstening to events
that have large changes in y. I also ran an autoformatter that added
a ton of extra changes.